Implement authorization using Gates

+++ app/Http/Controllers/QuestionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\AskQuestionRequest;
 use App\Models\Question;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionController extends Controller
 {
@@ -61,6 +62,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
         return view('questions.edit')->with(['question' => $question]);
     }
 
@@ -73,6 +78,10 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with(['success' => 'Your question has been updated']);
     }
@@ -86,6 +95,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, 'Access denied');
+        }
+
         $question->delete();
         return redirect()->route('questions.index')->with(['success' => 'Your question has been deleted']);
     }
